adjusted conflict accorded by merging
